[#13] feat: Repository Protocol 추가
