+++ App/Donmani/Helper/Repository/UserRepository.swift
@@ -9,18 +9,6 @@ import DNetwork
 import ComposableArchitecture
 import Domain
 
-protocol UserRepository {
-    func registerUser() async throws -> User
-    func getUserName() -> String
-    func updateUserName(newUserName: String) async throws -> User
-    func postUpdateToken(token: String) async throws -> String
-    func putLastLogin() async throws
-    func putNoticeStatus() async throws
-    func getNoticeStatus() async throws -> Bool
-    func putRewardStatus() async throws
-    func getRewardStatus() async throws -> Bool
-}
-
 struct DefaultUserRepository: UserRepository {
     private let dataSource = UserAPI()
     private var keychainDataSource: KeychainDataSource
