--- Domain/Domain/Domain.swift
@@ -1,9 +1,0 @@
-//
-//  Domain.swift
-//  Domain
-//
-//  Created by 문종식 on 6/8/26.
-//
-
-import Foundation
-
